% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pageurls.R
\name{get_pageurls}
\alias{get_pageurls}
\title{Get PageUrls}
\usage{
get_pageurls(matomo_config = set_matomo_config())
}
\arguments{
\item{matomo_config}{matomo configuration (default: \code{\link{set_matomo_config}})}
}
\value{
data frame pageurls and with Matomo indicators
}
\description{
Get PageUrls
}
\seealso{
https://developer.matomo.org/api-reference/reporting-api
}
