% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_matomo_config.R
\name{set_matomo_config}
\alias{set_matomo_config}
\title{Set Matomo configuration}
\usage{
set_matomo_config(
  url = Sys.getenv("MATOMO_URL"),
  api_key = Sys.getenv("MATOMO_APIKEY"),
  site_id = Sys.getenv("MATOMO_SITEID"),
  format = "json",
  date_min = "2018-06-05",
  date_max = Sys.Date()
)
}
\arguments{
\item{url}{url to Matomo tracking (e.g. https://matomo.example.com/index.php), default: Sys.getenv("MATOMO_URL")}

\item{api_key}{Matomo tracking API key, default: Sys.getenv("MATOMO_APIKEY")}

\item{site_id}{site id, default: Sys.getenv("MATOMO_SITEID")}

\item{format}{export data format  (default: "json")}

\item{date_min}{first date (default: "2018-06-05", start of KWB-Matomo tracking)}

\item{date_max}{last date (default: Sys.Date())}
}
\value{
list with Matomo configuration
}
\description{
Set Matomo configuration
}
